LayoutTests/editing/execCommand/5573879.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/dump-as-markup.js"></script>
<div id="div" contenteditable="true"><ul><li>foo</li><li>bar</li></ul></div>
<script>

var div = document.getElementById("div");
div.focus();

document.execCommand("SelectAll");
document.execCommand("RemoveFormat");

Markup.description('This tests to make sure that RemoveFormat destroys lists if they are fully selected.  You should see foo<br>bar below.');
Markup.dump(div);

</script>
</body>
</html>
