LayoutTests/editing/execCommand/5573879-expected.txt

This tests to make sure that RemoveFormat destroys lists if they are fully selected.  You should see foo<br>bar below.
| <div>
|   "<#selection-anchor>foo"
|   <br>
|   "bar<#selection-focus>"
